Game state evaluation
Created isEndGame() function that checks if a game is in a stalemate or
checkmate by checking every possible move if they are valid or not. if
they are all invalid, then the game is over and it is either a checkmate
or a stalemate depending on whether or not the king is currently in check.